fix(ui): pin the Files sidebar header and scroll-fade its rows
WorkspaceFilesPanel wrapped FileTree's header (root label, refresh,
collapse) and its rows in one overflow-y-auto div, so the header scrolled
away with a long tree. FileTree now owns a SidebarHeader/SidebarScrollRegion
split internally (mirroring SessionSidebar), keeping the header pinned and
giving the rows the shared scroll-fade treatment instead of a hand-rolled
scroller. TreeNode and the tree's path/sort helpers move to their own files
to keep FileTree.tsx under the repo's line limit after the split.
